bonus

+++ src/components/ReduxForm/Promotions/PricesBonusProductField.js
@@ -1,6 +1,6 @@
 import _ from 'lodash'
 import React from 'react'
-import {Row, Col} from 'react-flexbox-grid'
+import {Row} from 'react-flexbox-grid'
 import {compose, withReducer, withHandlers} from 'recompose'
 import injectSheet from 'react-jss'
 import IconButton from 'material-ui/IconButton'
@@ -41,27 +41,17 @@ const enhance = compose(
             }
         },
         table: {
-            marginTop: '20px',
-            '& .row': {
-                margin: '0',
-                height: '40px',
-                '&:first-child': {
-                    fontWeight: '600'
-                },
-                '&:last-child:after': {
+            display: 'flex',
+            margin: '0 -30px'
+        },
+        halfTable: {
+            width: '50%',
+            padding: '20px 30px',
+            '& .dottedList': {
+                margin: '5px 0',
+                padding: '0',
+                '&:after': {
                     display: 'none'
-                },
-                '& > div': {
-                    textOverflow: 'ellipsis',
-                    whiteSpace: 'nowrap',
-                    padding: '0 8px',
-                    overflow: 'hidden',
-                    '&:first-child': {
-                        paddingLeft: '0'
-                    },
-                    '&:last-child': {
-                        paddingRight: '0'
-                    }
                 }
             }
         },
@@ -106,24 +96,23 @@ const enhance = compose(
             margin: '-20px -30px 0',
             backgroundColor: '#f1f5f8',
             position: 'relative',
-            '& > div:last-child': {
-                alignSelf: 'center',
-                textAlign: 'center',
-                width: '20%'
+            '& > div': {
+                textAlign: 'right',
+                width: '50%',
+                padding: '0 30px'
             },
             '& > div > div > div:first-child': {
                 overflow: 'hidden'
             }
         },
         bonusProduct: {
-            borderRight: '1px #ccc solid',
-            padding: '0 30px',
-            width: '40%'
+            borderRight: '1px #ccc solid'
         },
         giftProduct: {
             extend: 'bonusProduct'
         },
         subTitle: {
+            textAlign: 'left',
             fontWeight: 'bold',
             marginBottom: '5px'
         },
@@ -223,6 +212,10 @@ const PricesBonusProductField = ({classes, state, dispatch, handleAdd, handleRem
                                 />
                             </div>
                         </div>
+                        <FlatButton
+                            label="Добавить"
+                            labelStyle={{color: '#12aaeb'}}
+                            onTouchTap={handleAdd}/>
                     </div>
                     <div className={classes.giftProduct}>
                         <div className={classes.subTitle}>Товар в подарок</div>
@@ -253,8 +246,6 @@ const PricesBonusProductField = ({classes, state, dispatch, handleAdd, handleRem
                                 />
                             </div>
                         </div>
-                    </div>
-                    <div>
                         <FlatButton
                             label="Добавить"
                             labelStyle={{color: '#12aaeb'}}
@@ -265,34 +256,41 @@ const PricesBonusProductField = ({classes, state, dispatch, handleAdd, handleRem
             {error && <div className={classes.error}>{error}</div>}
             {!_.isEmpty(bonusProducts)
                 ? <div className={classes.table}>
-                    <div className={classes.subTitle}>Список бонусных товаров</div>
-                    <div>
-                        <Row className="dottedList">
-                            <Col style={{width: '34%'}}>Бонусный товар</Col>
-                            <Col style={{width: '11%'}}>Кол-во</Col>
-                            <Col style={{width: '34%'}}>Подарок</Col>
-                            <Col style={{width: '11%'}}>Кол-во</Col>
-                        </Row>
+                    <div className={classes.halfTable}>
+                        <div className={classes.subTitle}>При покупке:</div>
                         {_.map(bonusProducts, (item, index) => {
                             const bonusProduct = _.get(item, ['bonusProduct', 'value', 'name'])
                             const bonusMeasurement = _.get(item, ['bonusProduct', 'value', 'measurement', 'name'])
                             const bonusAmount = numberFormat(_.get(item, 'bonusAmount'), bonusMeasurement)
+                            if (_.isEmpty(bonusProduct)) {
+                                return false
+                            }
+                            return (
+                                <div key={index} className="dottedList">
+                                    <div>{bonusProduct} - {bonusAmount}</div>
+                                    <IconButton onTouchTap={() => handleRemove(index)}>
+                                        <DeleteIcon color="#666666"/>
+                                    </IconButton>
+                                </div>
+                            )
+                        })}
+                    </div>
+                    <div className={classes.halfTable}>
+                        <div className={classes.subTitle}>Клиент получает в подарок:</div>
+                        {_.map(bonusProducts, (item, index) => {
                             const giftProduct = _.get(item, ['giftProduct', 'value', 'name'])
                             const giftMeasurement = _.get(item, ['giftProduct', 'value', 'measurement', 'name'])
                             const giftAmount = numberFormat(_.get(item, 'giftAmount'), giftMeasurement)
-
+                            if (_.isEmpty(giftProduct)) {
+                                return false
+                            }
                             return (
-                                <Row key={index} className="dottedList">
-                                    <Col style={{width: '34%'}}>{bonusProduct}</Col>
-                                    <Col style={{width: '11%'}}>{bonusAmount}</Col>
-                                    <Col style={{width: '34%'}}>{giftProduct}</Col>
-                                    <Col style={{width: '11%'}}>{giftAmount}</Col>
-                                    <Col style={{width: '10%'}}>
-                                        <IconButton onTouchTap={() => handleRemove(index)}>
-                                            <DeleteIcon color="#666666"/>
-                                        </IconButton>
-                                    </Col>
-                                </Row>
+                                <div key={index} className="dottedList">
+                                    <div>{giftProduct} - {giftAmount}</div>
+                                    <IconButton onTouchTap={() => handleRemove(index)}>
+                                        <DeleteIcon color="#666666"/>
+                                    </IconButton>
+                                </div>
                             )
                         })}
                     </div>
@@ -300,7 +298,8 @@ const PricesBonusProductField = ({classes, state, dispatch, handleAdd, handleRem
                 : <div className={classes.imagePlaceholder}>
                     <div style={{textAlign: 'center', color: '#adadad'}}>
                         <img src={Groceries} alt=""/>
-                        <div>Вы еще не выбрали ни одного товара. <br/> <a onClick={() => dispatch({open: !state.open})}>Добавить</a> товар?
+                        <div>Вы еще не выбрали ни одного товара. <br/> <a onClick={() => dispatch({open: !state.open})}>Добавить</a>
+                            товар?
                         </div>
                     </div>
                 </div>
